# backend/docs/swagger.yaml
basePath: /
definitions:
  internal_handler.chatEnergyResponse:
    properties:
      limit:
        type: integer
      mood_label:
        type: string
      remaining:
        type: integer
      used:
        type: integer
    type: object
  internal_handler.completeRegistrationRequest:
    properties:
      display_name:
        type: string
    type: object
  internal_handler.createAnonymousSessionRequest:
    properties:
      device_id:
        type: string
      timezone:
        description: |-
          IANA timezone (e.g. "Europe/Istanbul"), used to reset daily quotas
          at the user's local midnight. Optional; defaults to UTC. Ignored
          for a returning device_id — timezone is set once, at creation.
        type: string
    type: object
  internal_handler.createAnonymousSessionResponse:
    properties:
      access_token:
        type: string
      has_completed_onboarding:
        description: |-
          HasCompletedOnboarding lets the client decide whether to show
          onboarding without a separate round trip — see
          service.SessionTokens.HasCompletedOnboarding for why this can't
          just be a client-local flag.
        type: boolean
      is_anonymous:
        type: boolean
      refresh_token:
        type: string
      user_id:
        type: string
    type: object
...
  internal_handler.errorResponse:
    properties:
      code:
        $ref: '#/definitions/perchly-backend_internal_apierror.Code'
      error:
        type: string
    type: object
  internal_handler.logoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  internal_handler.refreshSessionRequest:
    properties:
      refresh_token:
        type: string
    type: object
  internal_handler.requestEmailCodeRequest:
    properties:
      email:
        type: string
    type: object
...
  internal_handler.sessionResponse:
    properties:
      access_token:
        type: string
      has_completed_onboarding:
        description: |-
          HasCompletedOnboarding lets the client decide whether to show
          onboarding without a separate round trip — see
          service.SessionTokens.HasCompletedOnboarding for why this can't
          just be a client-local flag.
        type: boolean
      refresh_token:
        type: string
    type: object
...
  internal_handler.verifyEmailCodeRequest:
    properties:
      access_token:
        description: |-
          AccessToken is the caller's current anonymous session's access
          token, if any — see AuthService.VerifyEmailCode's Durum A/B split.
        type: string
      code:
        type: string
      email:
        type: string
    type: object
  internal_handler.verifyEmailCodeResponse:
    properties:
      access_token:
        type: string
      has_completed_onboarding:
        description: |-
          HasCompletedOnboarding lets the client decide whether to show
          onboarding without a separate round trip — see
          service.SessionTokens.HasCompletedOnboarding for why this can't
          just be a client-local flag.
        type: boolean
      is_new_registration:
        type: boolean
      refresh_token:
        type: string
    type: object
  perchly-backend_internal_apierror.Code:
    enum:
    - invalid_request_body
    - unauthorized
    - internal_error
    - session_create_failed
    - invalid_email
    - email_code_send_failed
    - too_many_attempts
    - verification_code_expired
    - invalid_verification_code
    - verification_failed
    - invalid_display_name
    - user_not_found
    - profile_update_failed
    - invalid_refresh_token
    - session_refresh_failed
    - logout_failed
    - invalid_persona_id
    - persona_not_found
    - persona_fetch_failed
    - personas_list_failed
    - persona_not_age_appropriate
    - persona_trait_out_of_range
    - persona_traits_failed
    - invalid_conversation_id
    - conversation_not_found
    - conversation_forbidden
    - conversation_create_failed
    - conversations_list_failed
    - conversation_fetch_failed
    - conversation_context_missing
    - messages_list_failed
    - invalid_message_id
    - message_not_found
    - message_content_required
    - message_send_failed
    - streaming_unsupported
    - reaction_emoji_required
    - invalid_reaction_emoji
    - cannot_react_to_own_message
    - reaction_failed
    - invalid_age_range
    - invalid_mood_preference
    - invalid_selected_persona_id
    - preferred_name_required
    - preferred_name_invalid
    - onboarding_save_failed
    - onboarding_fetch_failed
    - onboarding_profile_not_found
    - quota_check_failed
    - quota_exceeded
    - chat_energy_failed
    type: string
    x-enum-varnames:
    - CodeInvalidRequestBody
    - CodeUnauthorized
    - CodeInternal
    - CodeSessionCreateFailed
    - CodeInvalidEmail
    - CodeEmailCodeSendFailed
    - CodeTooManyAttempts
    - CodeVerificationExpired
    - CodeInvalidCode
    - CodeVerificationFailed
    - CodeInvalidDisplayName
    - CodeUserNotFound
    - CodeProfileUpdateFailed
    - CodeInvalidRefreshToken
    - CodeSessionRefreshFailed
    - CodeLogoutFailed
    - CodeInvalidPersonaID
    - CodePersonaNotFound
    - CodePersonaFetchFailed
    - CodePersonasListFailed
    - CodePersonaNotAgeAppropriate
    - CodePersonaTraitOutOfRange
    - CodePersonaTraitsFailed
    - CodeInvalidConversationID
    - CodeConversationNotFound
    - CodeConversationForbidden
    - CodeConversationCreateFailed
    - CodeConversationsListFailed
    - CodeConversationFetchFailed
    - CodeConversationContextMissing
    - CodeMessagesListFailed
    - CodeInvalidMessageID
    - CodeMessageNotFound
    - CodeMessageContentRequired
    - CodeMessageSendFailed
    - CodeStreamingUnsupported
    - CodeReactionEmojiRequired
    - CodeInvalidReactionEmoji
    - CodeCannotReactToOwnMessage
    - CodeReactionFailed
    - CodeInvalidAgeRange
    - CodeInvalidMoodPreference
    - CodeInvalidSelectedPersonaID
    - CodePreferredNameRequired
    - CodePreferredNameInvalid
    - CodeOnboardingSaveFailed
    - CodeOnboardingFetchFailed
    - CodeOnboardingProfileNotFound
    - CodeQuotaCheckFailed
    - CodeQuotaExceeded
    - CodeChatEnergyFailed
...
info:
  contact: {}
  description: |-
    Perchly sohbet uygulamasının REST API'si — kimlik doğrulama, onboarding, persona, konuşma/mesajlaşma ve kota uçları.
    Kimlik doğrulama gerektiren uçlar için "Authorize" düğmesinden `Bearer <access_token>` girin.
  title: Perchly Backend API
  version: "1.0"
paths:
  /auth/anonymous:
    post:
      consumes:
      - application/json
      description: device_id'ye göre yeni bir anonim kullanıcı oluşturur veya var
        olanı döner. Gövde opsiyoneldir.
      parameters:
      - description: device_id ve opsiyonel IANA timezone
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handler.createAnonymousSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.createAnonymousSessionResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Anonim oturum oluştur
      tags:
      - auth
  /auth/email/request-code:
    post:
      consumes:
      - application/json
      description: E-postaya bir doğrulama kodu gönderir. Kayıtlı olup olmadığına
        bakılmaksızın her zaman aynı genel başarı cevabını döner.
      parameters:
      - description: email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.requestEmailCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: E-posta doğrulama kodu isteği
      tags:
      - auth
  /auth/email/verify-code:
    post:
      consumes:
      - application/json
      description: Kodu doğrular; access_token dolu gönderilirse mevcut anonim oturumu
        aynı user_id ile e-postalı hesaba yükseltir (Durum B), boşsa/olmayan bir e-postaysa
        yeni hesap açar (Durum A).
      parameters:
      - description: email, code, opsiyonel access_token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.verifyEmailCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.verifyEmailCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: E-posta doğrulama kodunu doğrula
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Verilen refresh token'ı iptal eder.
      parameters:
      - description: refresh_token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.logoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Çıkış yap
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh token ile yeni bir access/refresh token çifti üretir.
      parameters:
      - description: refresh_token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.refreshSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.sessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Oturumu yenile
      tags:
      - auth
  /auth/register/complete:
    post:
      consumes:
      - application/json
      description: users.display_name'i yalnızca bu uç yazar — onboarding'deki preferred_name
        (hitap) ile karıştırılmaz.
      parameters:
      - description: display_name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.completeRegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Kayıt sonrası görünen adı belirle
      tags:
      - auth
...
  /users/chat-energy:
    get:
      description: 'Keşfet ekranındaki tek sayı: tüm aktif personalar arasında min(remaining)
        olarak hesaplanır. Salt okunur; gerçek mesaj gönderme uçlarındaki 429 davranışını
        etkilemez.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.chatEnergyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Günün sohbet enerjisi
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"